requirements.txt
python-dotenv
sqlalchemy
flask
